+++ NEWS.md
@@ -6,6 +6,8 @@ reghelper 0.3.0
 
 -   `graph_model` function extended to include `aov` and `glm` models.
 
+-   `build_model` function extended to include `aov` models.
+
 reghelper 0.2.0
 ===============